Initial chord progression implementation
Accepts list of chords we're interested, and prints out several
alternatives of how to play this progression. These are optimised for
"Ease of playability", i.e. how far one needs to move hand, how many
fingers need to move, and so on. Currently still mostly for guitar, but
we have some beginnings of Ukulele support as well, so we might re-visit
this in future.